backend/requirements.txt
Django==4.2.1
django-environ==0.9.0
django-extensions==3.2.1
psycopg2-binary==2.9.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-filters==1.0.0.dev2
django-storages==1.11.1
django-cors-headers==3.7.0
django-braces==1.14.0
arrow==1.1.0
gunicorn==20.1.0
pytz==2021.1
pandas==2.0.2
